initial implementation of sidebar
Summary:
this is an initial implementation of the sidebar, which just has links to the cluster/namespace view.
the profile menu is now moved to the bottom of the sidebar. for now, it still has the profile menu, until we figure out how we want to do the footer.

Test Plan:
screenshots:
{F86510}

{F86511}

updated screenshots:
{F86562}

{F86563}

{F86564}

Reviewers: zasgar, philkuz, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5501

GitOrigin-RevId: c12a24ff05538a23a8dde31be96c45605cd1364f