Use appBar + sidebar in admin page
Summary:
we added a sidebar to the live view page, but I realized I also need to add it to the admin page for consistency.
since the sidebar also consists of a little hamburger menu in the appbar, and i didnt want to copy & paste that styling into the admin page, i pulled out the common components (appbar + sidebar) into a different file. i wasn't sure what to call this, since it consists of both the top+side bar. for now i called it NavBars, but please let me know if you have any better naming ideas.

Test Plan:
screenshot:
{F86665}

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5513

GitOrigin-RevId: d64c2dd0bdd9d98bfdc2319aed444d9d852ec1b5