Metadata agent updates: dequeue by max # of bytes
Summary:
customer was running into an issue where their heartbeat ack messages were too large to send over nats. we had changed the max # of updates to be 100, which usually works for the most part. hwoever, it seems some of their updates are larger than expected.
instead of giving by the # of updates, we should just go by the size of the updates.

Test Plan: ran with a small MaxBytesToDequeue and made sure things were still running

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5514

GitOrigin-RevId: da61615200777320101179593c8dca2268b8e5ac