PC-436 Make editor resize with parent
Summary:
the editor currently doesn't resize when you resize the parent window. this is because the editor component requires an event to trigger a resize.
right now, it is listening to the window resize event. this is unfortunately not triggered when resizing an indvidual div. we can have the resizing action trigger a window resize event. also added a timer so we can debounce and not send out too many resize events while the user is actively resizing.

Test Plan: ran webpack and resized the drawer

Reviewers: zasgar, philkuz, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5532

GitOrigin-RevId: c3e9318bda53751684685cf5d3d8832e8418efbf