Initial changes to topbar styling
Summary:
here are some initial changes to the topbar styling. this will transform more once we have the breadcrumbs for the script args.
here are the changes made:
- topbar height update
- topbar color updates

for now, i've moved the script args in their current form out of the topbar. this will be replaced with breadcrumbs in a following diff.
the cluster selector will also be moved out of the topbar once the breadcrumbs are done.

just wanted to get the new color/height changes in so people can start looking at it in-app.

Test Plan:
screenshot:
{F86891}

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5528

GitOrigin-RevId: 9632b1abb54c429893b515ec1a40adda46a16b04