Dynamic Tracing: DataTableSchema
Summary:
Converts the OutputStruct into a DataTableSchema.

This turned out to be somewhat tricky, because DataTableSchema is fully constexpr, and only stores views into static data.

So to make the model work, needed to define DynamicDataTableSchema which is essentially a DataTableSchema plus the storage that holds the values that DataTableSchema refers to.

Test Plan: Test case added

Reviewers: yzhao, #engineering

Reviewed By: yzhao, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5526

GitOrigin-RevId: a488dd2522de63c046e8bcf87e56dcb7e6281211