PC-293: Store the statuses of Vizier control plane pods in the vzmgr db
Summary: TSIA, we want to store the statuses of Vizier control plane pods (but not pems or kelvins) in the cluster_info table so that when the admin UI inspects a cluster, it can find out the status of the non-query pods. We omit pems and kelvins 1. because they are already covered by the pxl get agent status query 2. because the stored message will scale too much on large clusters (whereas control plane pods won't scale linearly).

Test Plan: added unit tests

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, zasgar, #engineering

JIRA Issues: PC-293

Differential Revision: https://phab.corp.pixielabs.ai/D5016

GitOrigin-RevId: a8b4e1c6e3e0203ff43b963113f1c492142541a1