Remove email from auth signup flow
Summary: We don't need to get the email from the API since we can just request it using the Auth0 token. In the current flow we actually request it using the Auth0 API and compare it to the provided value. This seems unnecessary and just overcomplicates the API.

Test Plan: Existing tests updated.

Reviewers: michelle, nserrino, #engineering

Reviewed By: michelle, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5547

GitOrigin-RevId: 62d23447111135c6a4caf2b72c66286eb1714ec0