PC-439 move script errors from datadrawer to canvas
Summary:
we want move script execution/compilation errors out of the data drawer.
for now, we are just using the default MUI styling. we can tweak it as necessary.
the mock shows a button in the alert where we can close the alert, but given that there is no other way to see the error for now, i have left that out.

Test Plan:
screenshot:
{F87109}

Reviewers: zasgar, nserrino, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5541

GitOrigin-RevId: a162c531851dbc9f4ec7276d7c532f52c836f7b8