[PC-450] prefetch monaco instead of loading it with the rest of the chunks
Summary: Monaco was slow to load and we don't need it immediately, so we prefetch for now.

Test Plan: Tested locally with throttling and the page loads w/o loading monaco. Need some feedback on whether I'm actually not loading monaco as part of the chunk though.

Reviewers: michelle, jamesbartlett, zasgar, nserrino, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-450

Differential Revision: https://phab.corp.pixielabs.ai/D5554

GitOrigin-RevId: be252de9b8a5726c6cd23ba5426408910e020d36