PC-452, PC-448 fix colors of data tables/titles
Summary:
fix the colors of data tables/widget borders to match what they originally were.
also update titles to be white, to match the color on figma. the sizing on figma indicates 20px, which is what typography.subtitle1 in figma should be, so the font size remains the same

Test Plan:
screenshots:
{F87217}

{F87218}

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5557

GitOrigin-RevId: a1b60780ab27625afd41880b98c26f07b336ae4f