PC-433: Postgresql views (mirroring mysql views)
Summary:
TSIA, one of our prospective customers was interested in psql and didn't know we collected that data since we didn't have any views for it.
Note that it differs from the mysql views in minor ways. It's missing the status field, so we can't compute failures. Ideally we would have a common schema for postgres and mysql.

Test Plan: ran it

Reviewers: zasgar, philkuz, #engineering

Reviewed By: philkuz, #engineering

Subscribers: ishan

JIRA Issues: PC-433

Differential Revision: https://phab.corp.pixielabs.ai/D5545

GitOrigin-RevId: 90d5779c1e7110cf1de136573d404c11c3212746