Fix bug with reverse hover
Summary: There was a weird bug where if the line was completely flat the reverse hover wouldn't work. Tried in an unpatched version of vega and the bug still exists there so it must be a vega bug not an issue with the patches. This is a workaround.

Test Plan: Tried reverse hover for a completely flat line.

Reviewers: philkuz, #engineering

Reviewed By: philkuz, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5592

GitOrigin-RevId: 2845cbe4b800dd99f56e4c6315ee40670f9739e0