Fix bug where data drawer JSON was unstyled, refactor format-data classes, reactivate tests.
Summary:
Most of this diff is refactoring format-data to follow our guidelines using withStyles.
The unit tests for format data are added back in in the new format we want for those types of tests.
Also this diff removes a bunch of places where className is passed down across various components.
Finally it fixes a bug with @zasgar's fix for where the data drawer JSON was not displayed correctly when clicked.

Test Plan: unit

Reviewers: zasgar, michelle, #engineering

Reviewed By: zasgar, #engineering

Subscribers: zasgar

Differential Revision: https://phab.corp.pixielabs.ai/D5601

GitOrigin-RevId: 689ef76d009e8490061cebf31b1ffe52f70104b0