Unblock metadata agent registration handler
Summary:
We were seeing an issue in large test cluster where all agents would continuously crash once kelvin crashed.
this is what happened:
- agents are all running and heartbeating
- kelvin crashes
- kelvin reregisters
- metadata populates kelvin's update queue in the agent registration handler. this may take 30s because it hits etcd a few times to fetch pods/services/namespaces separately, so that updates are given in the correct order.
- metadata continues handling agent hb messages... but since 30s has passed, all of these agents have now died.
- metadata is getting a continuous stream of updates, all 30s behind, so is never able to catch up and recover.

to solve this, i moved the code for populating the update queue in a separate goroutine. this may mean that the agent's first hearbeat ack may not contain any udpates, but this is fine. the agent will eventually get all of the updates in a later heartbeat ack.

Test Plan: updated tests, tested on large test cluster

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5602

GitOrigin-RevId: 3966d27ff2081e71a6b97549bd8b4f1731736fe7