Reduce column sizing jank by seeding random sample
Summary:
We used to random sample different records each time the table was rendered. This lead to jank whenever we clicked on a quantile since the column would rerender.

We don't actually need it to be that random so just introduced a seeded random number generator.

Test Plan: Click around on quantiles. much better ...

Reviewers: nserrino, michelle, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5609

GitOrigin-RevId: ef784ef71c758187be478e965b09df6f136c2219