Fix cluster selector
Summary:
this useEffect call is only supposed to be called when the page first loads, so that we can parse the cluster in the URL and update it.
however, it was incorrectly getting called every time the deps changed and would keep switching the cluster back to the previous cluster.

Test Plan: ran webpack and switched clusters back and forth

Reviewers: zasgar, nserrino, #engineering

Reviewed By: nserrino, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5616

GitOrigin-RevId: c11febb4fa57359b63975e4b8b5939467fbf657c