Clang-10 and actually fix dev image to be Ubuntu 20.04.
Summary:
This fixes the following issue related to libtinfo on PEM/Kelvin start:
`libtinfo.so.5: cannot open shared object file: No such file or directory`

When we updated to 20.04, I think I changed the base image but not the active dev image.
This means that we were building all the images using the old ubuntu 18.04.
James noticed a break with skaffold that caused the PEMs to break because of relying on deprecated terminfo5 so he changed our sid deploy dependency to terminfo6. Unfortunately, this broke our deploy images.

I updated the dev images to be ubuntu 18.04, but sadly Clang on that relies on terminfo5 so it's broken with ubuntu 20.04. It's fixable by installing terminfo5, but this might break our code if it links to v5 instead of v6, since we only ship v6.

Updating the old clang install didn't work. clang-9 does not compile with clang-10. Ubuntu 20.04 ships with clang-10 so the compiler can't build our old clang-9 deb with terminfo6...

This updates the compiler to clang-10 with all the related changes necessary to make sure the code still compiles.￼

Test Plan:
`bazel test //...` in the docker VM.
`bazel run src/vizier/services/agent:pem_image` in the docker VM.

Reviewers: oazizi, michelle, jamesbartlett, philkuz, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5623

GitOrigin-RevId: a26680e32211f9aec5e42c1349c4cd35015e2213