PerfBufferSpec::name changed to std::string
Summary:
Decide that to put the SocketTraceConnector::kPerfBufferSpecs definition to the .cc file.

That way, I don't need to declare std::array<..., 4> kPerfBufferSpecs inside SocketTraceConnector,
which would have to be updated when new perf buffer is added.

Remove static declaration of callback functions as well, which are not used anywhere else.
That trickles the need to make SocketTraceConnector::Accept*() public member functions.

Test Plan: Jenkins

Reviewers: oazizi, #engineering

Reviewed By: oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5630

GitOrigin-RevId: 1d7954e3e34b074294b8ee855939efcebed343c1