Update metadata cache to use rwLock
Summary:
this is part of the effort to improve metadata registration/heartbeat handling speed.
looking at customer's cluster, i noticed that some of the Get functions in the cache were waiting for .3s to obtain the lock.
some of the cache functions, such as GetAll or GetAllWithPrefix requires a range request to etcd, and these results are then merged with the cache. range requests to etcd can be a little slow, and it doesn't make sense to block other read requests on the etcd call.
switched to read locks to speed up these get requests.

Test Plan: ran skaffold, made sure things still work as expected

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5631

GitOrigin-RevId: 37b42a92f4b7df4246f41781f2ec8c7c5c9fde93