PC-476: Show live view breadcrumbs when live view is trying to connect to cluster
Summary: /live redirects to the previously used cluster. When that cluster has become unhealthy, the user has no obvious way of getting out of the "connecting to cluster" state. This diff adds the live view breadcrumbs in the connecting to cluster scenario so the user can still see available clusters and switch to them if the current one is unhealthy.

Test Plan: ran it in staging which has all unhealthy or disconnected clusters.

Reviewers: michelle, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-476

Differential Revision: https://phab.corp.pixielabs.ai/D5636

GitOrigin-RevId: 836d49b1b88b0fe5c8702c9ddb7e56e687a32705