Use multiple locks in metadata cache
Summary:
a previous diff changed the lock in the cache to be a RWMutex so that Get calls don't get blocked when other Get calls make requests to etcd.
This helped speed up the Get calls in the metadata RegisterAgent/UpdateHeartbeat functions. However, after checking the function timing, Set calls were still taking a lot of time. This is because they still had to wait for those etcd calls... Which is ridiculous since the Set calls themselves only update a value in the in-memory cache.
To solve this, i added another lock. One lock is for locking the cache, and another is for locking updates to the underlying datastore (etcd in this case).
Get calls will acquire a read lock on both, but relinquish the cache lock earlier so that any Set calls can grab the lock if necessary.

I thought through this for a long time to make sure this won't hit deadlock. since we are always requiring the cacheLock to be held before grabbing the dsLock, it's not possible for one thread to be holding onto the cacheLock and waiting for the dsLock, while another thread holds onto the dsLock and waits for the cacheLock.

Test Plan: ran in skaffold and customer...

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5637

GitOrigin-RevId: 646a7dfe96caa3a1de6836f9878bd9774e8dc3ee