Add proto message for storing Probe Info in etcd
Summary:
This adds the proto format for how we plan to store ProbeInfo in etcd in MDS.
Currently, MDS stores protos from agentpb and metadatapb, which are all part of shared directories and used in a couple other places.
ProbeInfo is the only proto which is used internally in MDS, so I have created a new file for such protos.

the probe message just contains the ID, and the list of agents who have failed/suceeded to register the probe.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5671

GitOrigin-RevId: 4dcb89c1d192c3a76f2de5864ab66cba438e2c50