Add agentID to RegisterProbeResponse/ProbeInfoUpdate messages
Summary:
since these messages are just sent along the regular metadata update NATS channel, we need to include the agent ID so that we know which agent sent the message.
we also need to store the program in the ProbeInfo message in etcd, so that we know what probes to register if a new agent connects.

i also noticed that the RegisterProbeResponse is basically the same as the ProbeInfoUpdate message.
I am open to consolidating these two into one, with an additional "registered" boolean field. if a status OK is sent with registered as false, then that would mean the agent has just received the registerProbe request with no errors.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5699

GitOrigin-RevId: b9feca2d1c488fc77fa9f405cb92f4f792c2900e