UI: Add retry behavior for mutations
Summary:
when a script with a mutation is executed, we should retry the script until the mutation is ready or fails.
to do so, when a script is executed, we check if it contains any strings that indicate there is a mutation. if so, we set the mutation flag to "true" in the executeSciptRequest.

this diff assumes that if the mutation flag is set to true, then there will always be a MutationInfo included in the stream.
If the mutationInfo indicates that the status has grpc code UNAVAILABLE, then it is still in the pending process and we should wait for 5 seconds and retry script execution.
otherwise, if the grpc code is not UNAVAILABLE, it has either succeeded, in which there are query results which we can display, or it has failed.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5733

GitOrigin-RevId: 4e5029c0472e617527ed2e657f725b57a8410cc0