Add function to watch etcd keys with prefix
Summary:
we want to be able to watch when a tracepoint's lease expires in etcd.
since we are planning to use etcd's lease TTLs for this, we need to be notified when the key is deleted from etcd.
this diff adds a function which is a wrapper ontop of etcd's client watcher.

the next update should update metadatastore to have a "watchTracepoints" function which abstracts these events/keys into something the tracepoint manager can actually understand.

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5755

GitOrigin-RevId: dbe68be33b64fdd9ad20d3c9d8672ccc03ff95c1