PC-489: by default, sort UI tables by the first column in ascending order for visual consistency.
Summary:
It can be jarring to the user that the rows of tables come in a different order every time the query is executed.
This diff addresses some of that problem by default sorting ascending by the first column, which should fit most of our needs decently at least for now.
Time will be in ascending order as before and categorical things will go from a to z.

Test Plan: ran the ui.

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PC-489

Differential Revision: https://phab.corp.pixielabs.ai/D5740

GitOrigin-RevId: ee23c8bd90ff86c914a3391c78d715cc825d28f0