MDS: add functions for writing directly to etcd
Summary:
all of our functions currently write to the cache, and have it flushed out to etcd in the next 30s.
however, we now have some keys which we want to write directly to etcd (such as the lease key for tracepoints).
this diff adds those functions which can be used to write directly to the underlying datastore, rather than to the cache.

Test Plan: unit tests

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5758

GitOrigin-RevId: cd2e44e22f21f8f3bd670c2c922f52c6381a8643