PC-490: Propagate start time of script if available when entity links are clicked
Summary: this adds the start time (if available) to entity links in table widgets, graph widgets, and request graph widgets. It does not add it to the nav bar links because there isn't a great way to do that cleanly. adding this because it was really annoying me to constantly reset the time on a customer cluster.

Test Plan: added unit test

Reviewers: michelle, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PC-490

Differential Revision: https://phab.corp.pixielabs.ai/D5762

GitOrigin-RevId: df6e1a1dd484fc2bcf2a1ef7d4ee460920ffd726