UpdateSchemas in metadata_store now computesSchema based on uuid
Summary: Part of adding schemas properly, we now actually computeSchemas. This manages a mapping of schemas to agents as part of what's stored in `/computedSchema`. If an agent removes a table, the computedSchema will remove that agent from the schema. If that was the only agent with the relation, the relation for that table is dropped entirely on the computedSchema.

Test Plan: add tests to show functionality

Reviewers: zasgar, nserrino, #engineering, michelle

Reviewed By: #engineering, michelle

Subscribers: michelle

Differential Revision: https://phab.corp.pixielabs.ai/D5790

GitOrigin-RevId: 2ccdffe2d24192b75a69165af5a87ab9e41dcdb7