PP-2057: Finish up metadata server side of GetAgentUpdates
Summary: GetAgentUpdates is now implemented as a server streaming grpc method on the metadata service. Next up will be to refactor query_broker to use it, then to remove the old endpoints that are no longer used by the query broker after that point from the metadata service.

Test Plan: added a beast of a test

Reviewers: michelle, philkuz, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2057

Differential Revision: https://phab.corp.pixielabs.ai/D5761

GitOrigin-RevId: 3c0aa7b4f75cff9a677348b685a56c8901c38481