Pass schema info as part of metadata
Summary:
50% hack, removed the need to do it on querybroker side.

Big thing here is that we used the info from UpdateSchemas() to now populate what's returned by GetSchemas().

Now `GetAgentTableMetadata` returns `repeated SchemaInfo` as part of the response, making the AgentInfo job much easier to manage. Still want to remove the dependence on other schema calls in the future, but this should enable us to properly broadcast queries to the correct targets.

When we get rid of `GetAgentTableMetadata` , we need to make sure the replacement assemebles SchemaInfo from ComputedSchema{} like we do in `convertToSchemaInfo`

Test Plan: added metadata tests

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5791

GitOrigin-RevId: 573c44514cb21c99e01b15844c201468e75489a3