Update arc unit to always check for pbgo file if it exists
Summary:
in D5445, the test that checks that pbgo files are up to date was modified, since there are some protos that do not have a pbgo file.
this has been leading to a lot of un-updated pbgo files.
updated the test so that if a pbgo file does exist in the directory, then we should alway ensure that its up to date.
this means non-go protos should not trigger an arc unit test failure for the pbgo test.

Test Plan: tried editing proto files and making sure that arc unit gave the correct result

Reviewers: oazizi, zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5807

GitOrigin-RevId: 8977bea754b669f35aaa28cb0ce7a63e95b2538a