PP-2050 - Query broker implementation of tracepoints.
Summary:
This adds implementation of tracepoints in the query broker.
More work is still needed to make this robust and certainly some cleanup work is needed
especially around the API. In the longer run we probably want to make this a fully streaming API.

Test Plan: Tested manually. We'll add more tests soon.

Reviewers: michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2050

Differential Revision: https://phab.corp.pixielabs.ai/D5815

GitOrigin-RevId: 1a0cd29818e2f1582b8cfc8cf4f4fd94f5760b5e