PP-2057: Improvements to GetAgentUpdate endpoint
Summary:
While working on the query broker side which calls metadata service's GetAgentUpdates, some limitations in the current implementation became more obvious. Packing the agent updates in an array by update type (deleted, data info, etc) causes us to lose the order of the updates which can create more complex logic to avoid bugs (since there is no global order and there is chunking, some cases include 1. what happens if you see an update for an agent that just got deleted, or 2. what if you receive an update to the data info for an agent whose create you haven't seen yet). Additionally it seems better to send an empty update on the update interval, rather than omit empty updates like we do today.

The solution instead is to make an update queue with a oneof field which simplifies the logic in both the client and server.

This diff updates the proto schema which is technically a breaking change but there are no clients using GetAgentUpdate today.

Test Plan: edited existing

Reviewers: michelle, philkuz, zasgar, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PP-2057

Differential Revision: https://phab.corp.pixielabs.ai/D5808

GitOrigin-RevId: 1433b482229bdc8aef4faa5eb57fd280831a07e5