Fix ASAN issue related to cidrs
Summary:
Fixes and issue where a cidr update can cause mds to get invalidated.
The current MDS state is assumed to be immutable but the CIDR code was violating this assumption. This
change defers the writes until the update period.

Also mark the state as const and added a comment to clarify.

Bug shown here: https://phab.corp.pixielabs.ai/P48

Test Plan: Added a test.

Reviewers: michelle, oazizi, #engineering

Reviewed By: michelle, oazizi, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5827

GitOrigin-RevId: bd5ef86b5f6c2bd82c5b8aa75c79fbb156bf726f