Creating tracepoint with name of existing, terminated tracepoint should work
Summary:
not sure if this is exactly the bug that was run into with the TTL, since I haven't gotten everything setup for tracepoints to actually work yet.
but this is definitely a case that needs to be handled.
if the tracepoint with the same name is already terminating, then we can create the new tracepoint with that name without trying to terminate the old tracepoint again.

Test Plan: unit test

Reviewers: zasgar, #engineering, philkuz

Reviewed By: #engineering, philkuz

Subscribers: philkuz

Differential Revision: https://phab.corp.pixielabs.ai/D5829

GitOrigin-RevId: 6c0eaa674e09d6a25c469c67134d079b0e67f258