PP-2071: Unify pod/container state/message/reason into one struct type to provide more informative icons in Live View
Summary: We currently make a status icon for pod and container phase/state. However, knowing that a pod is terminated does not tell you if it was for a healthy or unhealthy reason, you need to go into status/reason for that. This diff merges the 3 different udfs for pod/container into one that provides all 3 values as a struct.

Test Plan: edited

Reviewers: philkuz, zasgar, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2071

Differential Revision: https://phab.corp.pixielabs.ai/D5837

GitOrigin-RevId: 1e352158c65fb3f78c6141f074f57c29ade6cc90