TracepointMgr::Monitor set failed to terminated if expected status is terminated
Summary:
I was trying to deploy a tracepoint with a binary that didnt exist, and seeing if I could delete it.
because the tracepoint state never changed to terminated in TracepointManager::Monitor, no state update was getting sent out.
It turns out this case was actually hitting the default condition which just sets current_state = statuspb::FAILED_STATE.

Test Plan: ran in skaffold

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5840

GitOrigin-RevId: 6b3927bd1b7f3e598ea6b11049fa6f8b1719e0c7