[PC-494] clean up time series alignment logic and fix issue
Summary:
listeners were not added at the right point because we waited until currentView was not-nil -> this meant we didn't get initial values for the views.

Also switched from doing a hacky domain thing (2 time scales, domainRaw etc) to using an extent of the time_ column and passing that signal as the domain. Much cleaner and uses more optimizations in Vega.

Test Plan: Tested on a setup where http data was missing and alignes stuff properly

Reviewers: jamesbartlett, nserrino, michelle, #engineering

Reviewed By: jamesbartlett, #engineering

JIRA Issues: PC-494

Differential Revision: https://phab.corp.pixielabs.ai/D5861

GitOrigin-RevId: f289d7c68f25b2104956a928c3f0bce0b3bd848e