PC-480: Automatically select the first data table in the data drawer when it is open and there is no matching table name for the previous selection
Summary: If you have the data drawer open with tables [table1, table2] and table1 selected, then you switch live views to one that has [tableA, tableB], nothing will be selected in the open data drawer. This diff causes the automatic selection of "tableA" in that circumstance. If table1 was available in the new live view, it would sitll be selected.

Test Plan: ran it in various cases

Reviewers: philkuz, michelle, #engineering

Reviewed By: michelle, #engineering

JIRA Issues: PC-480

Differential Revision: https://phab.corp.pixielabs.ai/D5871

GitOrigin-RevId: 98b1c1cc3bdcbc7ffebfec15bd33ca41b67e62e4