PP-2102: Allow probes to be called from within exec funcs.
Summary: Previously, probes would only be invoked by a query as a mutation if they were in the main body of the script. Now, if they are wrapped in a function, and invoked via an exec arg (a pattern that works for queries), they will also be triggered. So now both probes and normal queries can be written in an exec funcs format. This will allow us to pass variables into probes from the UI.

Test Plan: added unit tests

Reviewers: philkuz, michelle, zasgar, #engineering

Reviewed By: philkuz, #engineering

JIRA Issues: PP-2102

Differential Revision: https://phab.corp.pixielabs.ai/D5873

GitOrigin-RevId: d2f57b7328d4f019857c7a1b6f11f9beae65cffc