Add vzwatcher interface
Summary:
this lays out how the vzwatcher API should look like. the vzwatcher is responsible for listening to the VizierConnected nats channel, transforming the VizierConnected proto messages
into a more digestible format, and forwarding the messages to another channel so that it can be consumed downstream.

Test Plan: n/a

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5868

GitOrigin-RevId: 3daa8d6a117317b3fd255cdc722622ebd11d259e