PP-2099: Fix join bug where compiler column pruning resulted in exception during execution
Summary:
PP-2099 shows a case where a join was causing kelvin to crash during execution. This was due to the fact that when joins were written, column pruning was not introduced into the compiler yet. As a result there was an assumption baked into the equijoin node execution code that both the build table and the probe table would contribute output columns. This caused an error in a script I ran where the join was only being used to filter rows down to ones matching the join keys, but the output columns from the build table were not actually used anywhere.

This fix stores the number of matching rows in the build buffer for a given join key. Previously, during the probe phase, the number of matching rows for the build buffer was computed from the length of the stored column chunks for that particular set of keys. This doesn't work if there are no output columns from the build buffer, so there is no index 0 column chunk to get the length from.

Test Plan: added tests to make sure that joins work successfully with 1. no columns from build buffer table 2. no columns from probe table.

Reviewers: zasgar, michelle, philkuz, jamesbartlett, #engineering

Reviewed By: zasgar, #engineering

JIRA Issues: PP-2099

Differential Revision: https://phab.corp.pixielabs.ai/D5863

GitOrigin-RevId: c4c8f59209dc918e0a8e322afd937b5d0f0a9c0e