Push public/private vizier images
Summary:
we want our rc images to be private only, and our actual releases to be in a public registry.
this diff updates the bazel rules so that we  can do that by specifying `--define public=True`.

unfortunately, there is some repetition because config settings are not evaluated in macros, which we are using to build the YAML files.
instead, we need to build both the private/public versions of the YAMLs, and use select to determine which YAML version is used

Test Plan: ran a staging build, and verified that works. havent run a prod (public) build, because we're not ready for an actual release yet. however, I did run the build rules locally to verify that the correct yamls get built.

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5890

GitOrigin-RevId: d96b4835ec5e47f51898f8169e6c9314a627abf2