Update vzmgr GetViziersByShard to take a a range of shard IDs
Summary:
the GetViziersByShard rpc call only takes a single shardID. however, we currently only have single replicas, which currently need to support all shardIDs.
looping through all of the possible shardIDs is impractical, so I updated the GetViziersByShard call to take a range of shards.
this will help us in the future even when we do have replicas, which handle a range of shards.
also updated the VizierInfo in the response to return more information that is needed by the vizierwatcher.

Test Plan: update unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5895

GitOrigin-RevId: 841da8e66811fa5a0b18fa97603f66d49e65b539