PC-499 Fix table rendering of JSONs in firefox
Summary:
firefox wasn't rendering the JSON rows in our tables the same way as chrome was.
there's a `white-space:nowrap` rule in the parent JSON div. this is respected as expected in chrome, where it puts everything on a single line.
but, for some reason, firefox seems to still break on any whitespace. to fix this, we need to use non-breaking spaces (&nsbp;). however, these don't render properly in jsx, so we use \u00A0 instead: https://stackoverflow.com/questions/37909134/nbsp-jsx-not-working

Test Plan: tried it in firefox

Reviewers: zasgar, nserrino, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5898

GitOrigin-RevId: bf842607996384bda2ce1fc2e0026e830b91c003