Update metadata reader to send updates to indexer in cluster UID-specific channel
Summary:
the metadata reader was checked in before our K8s UID changes were in.
before it was sending all of the updates to a generic "MetadataIndexer" channel.
instead, we should actually send it to a cluster UID-specific channel so that the indexer can pick it up and correctly index our updates by cluster UID.

I was thinking about switching metadata reader to use the new vzWatcher, since it basically does the same thing. however, the metadata reader is located within the vzmgr service itself, whereas the vzwatcher makes a rpc call to the vzmgr service (intended to be used outside of vzmgr).

Test Plan: unit test

Reviewers: zasgar, #engineering

Reviewed By: zasgar, #engineering

Differential Revision: https://phab.corp.pixielabs.ai/D5902

GitOrigin-RevId: 65df5133498a267c18debbe195f5202072b9041b